Python/project/requirements.txt
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
asyncpg==0.30.0
atlastk==0.13.4
attrs==25.3.0
backcall==0.2.0
beautifulsoup4==4.13.4
black==24.10.0
bleach==6.2.0
CacheControl==0.14.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
decorator==5.2.1
defusedxml==0.7.1
dnspython==2.7.0
docopt==0.6.2
email_validator==2.2.0
executing==2.2.0
fastapi==0.115.12
fastjsonschema==2.21.1
firebase-admin==6.7.0
flake8==7.1.1
google-api-core==2.24.2
google-api-python-client==2.167.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.2
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.72.0rc1
grpcio-status==1.72.0rc1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.8
httplib2==0.22.0
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
ipython==8.12.3
iso8601==2.1.0
isort==5.13.2
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mistune==3.1.3
msgpack==1.1.0
mypy-extensions==1.0.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
packaging==24.2
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
phonenumbers==8.13.53
pickleshare==0.7.5
pip-check==3.1
pipreqs==0.4.13
platformdirs==4.3.7
pluggy==1.5.0
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==6.30.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.10.6
pydantic-extra-types==2.10.1
pydantic-settings==2.6.0
pydantic_core==2.27.2
pyflakes==3.2.0
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.3
pypika-tortoise==0.3.2
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pyzmq==26.4.0
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
starlette==0.46.2
terminaltables==3.1.10
tinycss2==1.4.0
tornado==6.4.2
tortoise==0.1.1
tortoise-orm==0.23.0
traitlets==5.14.3
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.0
wcwidth==0.2.13
webencodings==0.5.1
yarg==0.1.9
pillow==11.2.1
python-multipart==0.0.20
